+++ src/api/all_params.rs
@@ -63,6 +63,8 @@ pub struct AllParameters {
     pub param_has_geo_info: Option<bool>,
     #[serde(rename = "hx")]
     pub param_has_extended_info: Option<bool>,
+    #[serde(rename = "hs")]
+    pub param_is_https: Option<bool>,
     #[serde(rename = "of")]
     pub param_offset: u32,
     #[serde(rename = "li")]